refactor: ファイル名に含まれる空白を削除
